[silly] add flag to only cond on first token
musicgen seems to prepend the conditioning, instead of adding to every token like uberduck tacotron multispeaker
against: they have attention so the model can see it all the time anyway
against: can't do mid-song cond switches because it's only used to it at the start
for: technically it'll be trained to be used to it at whenever infctx chunks start, and maybe the sparsity will make it more likely to accept a change?
for: maybe taco multispeaker sucked specifically because of addition? unsure